+++ src/app/layout.tsx
@@ -1,4 +1,5 @@
-import './globals.css'
+import '../styles/globals.css'
+import '../styles/client.css'
 
 export const metadata = {
   title: 'Welcome to guevaratorres.space!',
