+++ src/styles/client.css
@@ -96,7 +96,7 @@
 }
 
 body {  
-  background-image: url(./background.PNG);
+  background-image: url(../components/Project-Images/background.png);
   overflow: initial;
   background-color: bisque;
   background-size: cover;
